+++ packages/file-types/ome-tiff/src/OmeTiffAsObsSegmentationsLoader.js
@@ -1,10 +1,14 @@
 import { viv } from '@vitessce/gl';
 import { LoaderResult } from '@vitessce/vit-s';
-import { initializeRasterLayersAndChannels } from '@vitessce/spatial-utils';
+import {
+  initializeRasterLayersAndChannels,
+  coordinateTransformationsToMatrix,
+} from '@vitessce/spatial-utils';
 import OmeTiffLoader from './OmeTiffLoader';
 
-export default class RasterJsonAsObsSegmentationsLoader extends OmeTiffLoader {
+export default class OmeTiffAsObsSegmentationsLoader extends OmeTiffLoader {
   async load() {
+    const { coordinateTransformations } = this.options || {};
     const offsets = await this.loadOffsets();
     const { url, requestInit } = this;
 
@@ -18,6 +22,11 @@ export default class RasterJsonAsObsSegmentationsLoader extends OmeTiffLoader {
       // This load() method is the same as in ./OmeTiffLoader except we specify isBitmask here:
       metadata: {
         isBitmask: true,
+        ...(coordinateTransformations ? {
+          transform: {
+            matrix: coordinateTransformationsToMatrix(coordinateTransformations),
+          },
+        } : {}),
       },
     };