+++ packages/vit-s/src/index.js
@@ -82,6 +82,7 @@ export {
   obsSetsCsvSchema,
   anndataZarrSchema,
   imageOmeTiffSchema,
+  imageOmeZarrSchema,
   emptySchema,
 } from './file-options-schemas';
 export {